+++ example/lib/model/converter/date_time_converter.dart
@@ -1,0 +1,17 @@
+import 'package:json_annotation/json_annotation.dart';
+
+class DateTimeConverter implements JsonConverter<DateTime, String> {
+  const DateTimeConverter();
+
+  @override
+  DateTime fromJson(String json) {
+    if (json.contains(".")) {
+      json = json.substring(0, json.length - 1);
+    }
+
+    return DateTime.parse(json);
+  }
+
+  @override
+  String toJson(DateTime json) => json.toIso8601String();
+}