+++ example/lib/model/ogm.g.dart
@@ -24,6 +24,8 @@ OGM _$OGMFromJson(Map<String, dynamic> json) {
     someThinGHuGE: json['some_ThinG_huGE'] as String,
     securityRole: json['securityIndicator'] as String,
     mutableProperty: json['mutableProperty'] as String,
+    dateChange:
+        const DateTimeConverter().fromJson(json['dateChange'] as String),
   );
 }
 
@@ -36,4 +38,5 @@ Map<String, dynamic> _$OGMToJson(OGM instance) => <String, dynamic>{
       'some_ThinG_huGE': instance.someThinGHuGE,
       'securityIndicator': instance.securityRole,
       'mutableProperty': instance.mutableProperty,
+      'dateChange': const DateTimeConverter().toJson(instance.dateChange),
     };